Task2_NER_CV/requirements.txt
datasets==3.3.2
Pillow==11.1.0
torch==2.3.1
torchvision==0.18.1a0
transformers==4.49.0
